requirements.txt
asgiref==3.6.0
black==22.12.0
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.0
Django==4.1.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
flake8==6.0.0
isort==5.11.4
Markdown==3.4.1
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.10.3
platformdirs==2.6.2
psycopg2==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
python-dotenv==0.21.0
pytz==2022.7
sqlparse==0.4.3
tzdata==2022.7
